refactor: `RolloutGatewayMixin`
