+++ verifiers/envs/experimental/__init__.py
@@ -1,3 +1,4 @@
+from verifiers.envs.experimental.rollout_gateway_mixin import RolloutGatewayMixin
 from verifiers.envs.experimental.sandbox_mixin import SandboxMixin
 
-__all__ = ["SandboxMixin"]
+__all__ = ["RolloutGatewayMixin", "SandboxMixin"]
