+++ verifiers/envs/experimental/cli_agent_env.py
@@ -84,8 +84,6 @@ class CliAgentEnv(SandboxMixin, vf.MultiTurnEnv):
         )
         self.run_command = run_command
         self.poll_interval = poll_interval
-        self.interception_port = interception_port
-        self.interception_url = interception_url
         self.timeout_seconds = timeout_seconds
         self.docker_image = docker_image
         self.start_command = start_command
@@ -99,7 +97,16 @@ class CliAgentEnv(SandboxMixin, vf.MultiTurnEnv):
         self.advanced_configs = advanced_configs
         self.labels = labels
 
-        # Tunnel and interception server
+        self.init_interception(interception_port, interception_url)
+
+    def init_interception(
+        self,
+        interception_port: int = 8765,
+        interception_url: str | None = None,
+    ):
+        """Initialize interception server and tunnel resources. Call from __init__."""
+        self.interception_port = interception_port
+        self.interception_url = interception_url
         self._tunnel: Tunnel | None = None
         self._tunnel_lock = asyncio.Lock()
         self._interception_server = InterceptionServer(port=interception_port)
@@ -411,6 +418,8 @@ class CliAgentEnv(SandboxMixin, vf.MultiTurnEnv):
     @vf.teardown
     async def teardown_resources(self):
         """Stop Prime Tunnel and HTTP interception server."""
+        if not hasattr(self, "_tunnel_lock"):
+            return
         async with self._tunnel_lock:
             if self._tunnel is not None:
                 try:
@@ -437,7 +446,7 @@ class CliAgentEnv(SandboxMixin, vf.MultiTurnEnv):
         state.pop("background_job", None)
 
         rollout_id = state.get("rollout_id")
-        if rollout_id:
+        if rollout_id and hasattr(self, "_interception_server"):
             self._interception_server.unregister_rollout(rollout_id)
 
     @vf.stop
